dockers/video-handler/requirements.txt
aio_pika==6.8.0
pyyaml==5.4.1
nest_asyncio==1.5.1
validators==0.18.2